backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
redis[hiredio]==5.1.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
